manager 选择一个VIN 查看详情 有维修记录 和销售记录

+++ src/main/java/com/cs6400/carshop/bean/Transaction.java
@@ -12,7 +12,7 @@ import java.math.BigDecimal;
 public class Transaction {
     private Integer transaction_id;
     private String purchase_date;
-    private Integer customer_id;
+    private Long customer_id;
     private BigDecimal sold_price;
     private String sales_person_user_name;
     private String VIN;
